+++ UpgradeWorld/actions/SetWorldVersion.cs
@@ -1,12 +1,10 @@
 namespace UpgradeWorld;
-public class WorldVersion : ExecutedOperation {
-  readonly int Version = 0;
-  public WorldVersion(Terminal context, int version, bool start) : base(context, start) {
-    Version = version;
-  }
+public class WorldVersion(Terminal context, int version, bool start) : ExecutedOperation(context, start)
+{
+  readonly int Version = version;
 
-
-  protected override bool OnExecute() {
+  protected override bool OnExecute()
+  {
     var world = WorldGenerator.instance.m_world;
     world.m_worldGenVersion = Version;
     WorldGenerator.Initialize(world);
@@ -15,7 +13,8 @@ public class WorldVersion : ExecutedOperation {
     return true;
   }
 
-  protected override string OnInit() {
+  protected override string OnInit()
+  {
     return $"Updating world version to {Version}. This may change biome distribution.";
   }
 }